backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pydantic
requests
yt-dlp
youtube-transcript-api
playwright
